CALS-5654: client result and abstract authorizers refactored to eliminate code duplication
